Tidying up some things
