+++ Core/automation/lib/python/configuration.py
@@ -1,0 +1,2 @@
+openhabHost = "localhost"
+openhabPort = "8080"
