--- automation/jsr223/000_components/100_ExampleExtensionProvider.py
@@ -1,8 +1,0 @@
-from openhab import JythonExtensionProvider
-
-JythonExtensionProvider.addValue("Foo", 1)
-JythonExtensionProvider.addValue("Bar", "x")
-JythonExtensionProvider.addValue("Baz", {"a": 10})
-JythonExtensionProvider.addPreset("P1", ["Foo"], True)
-JythonExtensionProvider.addPreset("P2", ["Bar", "Baz"])
-
