--- automation/lib/python/esper/example0.py
@@ -1,4 +1,0 @@
-import esper.java
-
-from com.espertech.esper.client import EPServiceProviderManager
-print EPServiceProviderManager
