--- automation/lib/python/esper/java/log4j.properties
@@ -1,4 +1,0 @@
-log4j.rootLogger=debug, stdout
-log4j.appender.stdout=org.apache.log4j.ConsoleAppender
-log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
-log4j.appender.stdout.layout.ConversionPattern=%d{ISO8601} [%t] %-5p %c %x - %m%n
